requirements.txt
appdirs==1.4.4
Babel==2.13.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.1
frozendict==2.3.8
ghp-import==2.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-material-extensions==1.3
mkdocs-print-site-plugin==2.3.6
mkdocs-table-reader-plugin==2.0.3
mplfinance==0.12.10b0
multitasking==0.0.11
numpy==1.26.2
packaging==23.2
paginate==0.5.6
pandas==2.1.3
pathspec==0.11.2
peewee==3.17.0
Pillow==10.1.0
platformdirs==4.0.0
Pygments==2.16.1
pymdown-extensions==10.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tzdata==2023.3
urllib3==2.1.0
watchdog==3.0.0
webencodings==0.5.1
yfinance==0.2.31
